chore: cambio de tipo de variable en las migraciones

+++ Src/DTOs/GradeDto.cs
@@ -8,7 +8,7 @@ namespace GradesService.Src.DTOs.GradeDto{
         public string GradeName {get; set;} = string.Empty;
         [Required]
         [Range(1,7,ErrorMessage = "El valor de la calificación debe estar entre 1 y 7")]
-        public int GradeValue {get; set;}
+        public float GradeValue {get; set;}
         [Required]
         public string Comment {get; set;} = string.Empty;
         [Required]
